pyproject.toml
[project]
name = "peripheral-mcp"
version = "0.1.0"
description = "Public MCP server for The Peripheral OSINT data"
requires-python = ">=3.11"
dependencies = [
    "prefect>=3.0",
    "fastapi>=0.115",
    "fastmcp>=0.3",
    "httpx>=0.27",
    "uvicorn>=0.30",
    "supabase>=2.0",
    "python-dotenv>=1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project.scripts]
peripheral-mcp = "mcp.server:main"
peripheral-api = "api.main:app"

[tool.uv]
dev-dependencies = [
    "pytest>=8.0",
]
